Make extension relocatable
The supplied objects were already independent of the schema, we just
need to drop "db2." from all CREATE statements, and add some quoting in
places where the grammar would now report syntax errors.

The testsuite is now also schema-independent, except for the TRANSLATE()
test which needs an explicit schema for the reason already stated in the
README file, i.e. the name is already present in pg_catalog.

This also closes #4 which was caused by shell-only operator artifacts in
the public schema.